backend/requirements.txt
fastapi
uvicorn
python-multipart
docling
openai
pandas
pydantic
python-dotenv